# docs/openapi.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"example":{"status":"success","data":"Hello World!"}}}}},"tags":["App"]}},"/health":{"get":{"operationId":"HealthController_health","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","enum":["success"]},"data":{"$ref":"#/components/schemas/HealthDto"},"meta":{"type":"array"}}}}}}},"tags":["Health"]}},"/version":{"get":{"operationId":"VersionController_version","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","enum":["success"]},"data":{"$ref":"#/components/schemas/VersionDto"},"meta":{"type":"array"}}}}}}},"tags":["Version"]}},"/users":{"get":{"operationId":"UsersController_findAll","parameters":[{"name":"page","required":false,"in":"query","schema":{"type":"string"}},{"name":"size","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","enum":["success"]},"data":{"type":"array","items":{"$ref":"#/components/schemas/UserResponseDto"}},"meta":{"type":"object","$ref":"#/components/schemas/MetaResponseDto"}}}}}},"401":{"description":"You\"re not logged in. Missing/invalid JWT token."},"403":{"description":"You don\"t have enough permission!"}},"tags":["Users"]},"post":{"operationId":"UsersController_createUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUsersDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","enum":["success"]},"data":{"$ref":"#/components/schemas/UserResponseDto"},"meta":{"type":"array"}}}}}},"201":{"description":""},"401":{"description":"You\"re not logged in. Missing/invalid JWT token."},"403":{"description":"You don\"t have enough permission!"}},"tags":["Users"]}},"/users/{id}":{"get":{"operationId":"UsersController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","enum":["success"]},"data":{"$ref":"#/components/schemas/UserResponseDto"},"meta":{"type":"array"}}}}}},"401":{"description":"You\"re not logged in. Missing/invalid JWT token."},"403":{"description":"You don\"t have enough permission!"}},"tags":["Users"]},"patch":{"operationId":"UsersController_updateUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUsersDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","enum":["success"]},"data":{"$ref":"#/components/schemas/UserResponseDto"},"meta":{"type":"array"}}}}}},"401":{"description":"You\"re not logged in. Missing/invalid JWT token."},"403":{"description":"You don\"t have enough permission!"}},"tags":["Users"]},"delete":{"operationId":"UsersController_deleteUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""},"401":{"description":"You\"re not logged in. Missing/invalid JWT token."},"403":{"description":"You don\"t have enough permission!"},"default":{"description":"","content":{"application/json":{"example":null}}}},"tags":["Users"]}},"/products":{"get":{"operationId":"ProductsController_findAllProducts","parameters":[{"name":"page","required":false,"in":"query","schema":{"type":"string"}},{"name":"size","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","enum":["success"]},"data":{"type":"array","items":{"$ref":"#/components/schemas/ProductResponseDto"}},"meta":{"type":"object","$ref":"#/components/schemas/MetaResponseDto"}}}}}}},"tags":["Products"]},"post":{"operationId":"ProductsController_createProduct","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProductDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","enum":["success"]},"data":{"$ref":"#/components/schemas/ProductResponseDto"},"meta":{"type":"array"}}}}}},"201":{"description":""}},"tags":["Products"]}},"/products/{id}":{"get":{"operationId":"ProductsController_findProduct","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","enum":["success"]},"data":{"$ref":"#/components/schemas/ProductResponseDto"},"meta":{"type":"array"}}}}}}},"tags":["Products"]},"patch":{"operationId":"ProductsController_updateProduct","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProductDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","enum":["success"]},"data":{"$ref":"#/components/schemas/ProductResponseDto"},"meta":{"type":"array"}}}}}}},"tags":["Products"]},"delete":{"operationId":"ProductsController_deleteProduct","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""},"default":{"description":"","content":{"application/json":{"example":null}}}},"tags":["Products"]}},"/orders":{"get":{"operationId":"OrdersController_findAllOrders","parameters":[{"name":"page","required":false,"in":"query","schema":{"type":"string"}},{"name":"size","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","enum":["success"]},"data":{"type":"array","items":{"$ref":"#/components/schemas/OrderResponseDto"}},"meta":{"type":"object","$ref":"#/components/schemas/MetaResponseDto"}}}}}}},"tags":["Orders"]},"post":{"operationId":"OrdersController_createOrder","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOrderDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","enum":["success"]},"data":{"$ref":"#/components/schemas/OrderResponseDto"},"meta":{"type":"array"}}}}}},"201":{"description":""}},"tags":["Orders"]}},"/orders/{id}":{"get":{"operationId":"OrdersController_findOrder","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","enum":["success"]},"data":{"$ref":"#/components/schemas/OrderResponseDto"},"meta":{"type":"array"}}}}}}},"tags":["Orders"]},"patch":{"operationId":"OrdersController_updateOrder","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateOrderDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","enum":["success"]},"data":{"$ref":"#/components/schemas/OrderResponseDto"},"meta":{"type":"array"}}}}}}},"tags":["Orders"]},"delete":{"operationId":"OrdersController_deleteOrder","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""},"default":{"description":"","content":{"application/json":{"example":null}}}},"tags":["Orders"]}},"/auth/signup":{"post":{"operationId":"AuthController_signUp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUsersDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","enum":["success"]},"data":{"$ref":"#/components/schemas/AuthResponseDto"},"meta":{"type":"array"}}}}}},"201":{"description":""}},"tags":["Auth"]}},"/auth/signin":{"post":{"operationId":"AuthController_signIn","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","enum":["success"]},"data":{"$ref":"#/components/schemas/AuthResponseDto"},"meta":{"type":"array"}}}}}}},"tags":["Auth"]}},"/auth/logout":{"get":{"operationId":"AuthController_logout","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","enum":["success"]},"data":{"$ref":"#/components/schemas/LogoutResponseDto"},"meta":{"type":"array"}}}}}}},"tags":["Auth"]}},"/auth/refresh":{"get":{"operationId":"AuthController_refresh","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","enum":["success"]},"data":{"$ref":"#/components/schemas/RefreshResponseDto"},"meta":{"type":"array"}}}}}}},"tags":["Auth"]}}},"info":{"title":"Orders API","description":"The Orders API description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"MetaResponseDto":{"type":"object","properties":{"total":{"type":"number"},"lastPage":{"type":"number"},"currentPage":{"type":"number"},"totalPerPage":{"type":"number","nullable":true},"prevPage":{"type":"number","nullable":true},"nextpage":{"type":"number","nullable":true}},"required":["total","lastPage","currentPage","totalPerPage","prevPage","nextpage"]},"HealthDto":{"type":"object","properties":{"status":{"type":"string"},"timestamp":{"type":"string"}},"required":["status","timestamp"]},"VersionDto":{"type":"object","properties":{"app":{"type":"string"},"version":{"type":"string"},"env":{"type":"string"}},"required":["app","version","env"]},"UserResponseDto":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"email":{"type":"string"},"role":{"type":"string"},"createdAt":{"type":"string"},"updatedAt":{"type":"string"},"refresh_token":{"type":"string","nullable":true}},"required":["id","name","email","role","createdAt","updatedAt","refresh_token"]},"CreateUsersDto":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string","format":"email"},"password":{"type":"string","minLength":5},"passwordConfirm":{"type":"string","minLength":5},"refresh_token":{"type":"string","nullable":true}},"required":["name","email","password","passwordConfirm"]},"UpdateUsersDto":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string","format":"email"},"password":{"type":"string","minLength":5},"passwordConfirm":{"type":"string","minLength":5},"refresh_token":{"type":"string","nullable":true}}},"ProductResponseDto":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"price":{"type":"number"},"slug":{"type":"string"},"onStock":{"type":"boolean"},"createdAt":{"type":"string"},"updatedAt":{"type":"string"}},"required":["id","name","description","price","slug","onStock","createdAt","updatedAt"]},"CreateProductDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"price":{"type":"number"},"slug":{"type":"string"},"onStock":{"type":"boolean"}},"required":["name","description","price"]},"UpdateProductDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"price":{"type":"number"},"slug":{"type":"string"},"onStock":{"type":"boolean"}}},"OrderResponseDto":{"type":"object","properties":{"id":{"type":"string"},"status":{"type":"string"},"deliveryLocation":{"type":"string"},"createdAt":{"type":"string"},"userId":{"type":"string"}},"required":["id","status","deliveryLocation","createdAt","userId"]},"CreateOrderDto":{"type":"object","properties":{"status":{"type":"string","enum":["CREATED","AWAITING_PAYMENT","PAID","FULFILLED","COMPLETED","CANCELED","EXPIRED","REFUNDED"]},"deliveryLocation":{"type":"string"}},"required":["deliveryLocation"]},"UpdateOrderDto":{"type":"object","properties":{"status":{"type":"string","enum":["CREATED","AWAITING_PAYMENT","PAID","FULFILLED","COMPLETED","CANCELED","EXPIRED","REFUNDED"]},"deliveryLocation":{"type":"string"}}},"AuthResponseDto":{"type":"object","properties":{"userData":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"email":{"type":"string"},"role":{"type":"string"},"createdAt":{"type":"string"},"updatedAt":{"type":"string"}},"required":["id","name","email","role","createdAt","updatedAt"]},"tokens":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"}},"required":["accessToken","refreshToken"]}},"required":["userData","tokens"]},"AuthDto":{"type":"object","properties":{"email":{"type":"string","format":"email"},"password":{"type":"string"}},"required":["email","password"]},"LogoutResponseDto":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"email":{"type":"string"},"role":{"type":"string"},"createdAt":{"type":"string"},"updatedAt":{"type":"string"},"refresh_token":{"type":"string","example":null}},"required":["id","name","email","role","createdAt","updatedAt","refresh_token"]},"RefreshResponseDto":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"}},"required":["accessToken","refreshToken"]}}}}